engine: extract esm/layers/vo_table_layer.js
